+++ raspi/localization/probability_stat.cpp
@@ -5,12 +5,12 @@
 #include "probability_stat.h"
 
 namespace {
-  float StandardDistribution(float x) {
+  double StandardDistribution(double x) {
     return exp(- x * x / 2) / sqrt(2 * M_PI);
   }
 } // anonymous namespace
 
-ProbabilityStat::ProbabilityStat(float m, float v, float x) {
+ProbabilityStat::ProbabilityStat(double m, double v, double x) {
   mean_ = m;
   std_ = sqrt(v);
   prob_ = -1;
@@ -18,17 +18,17 @@ ProbabilityStat::ProbabilityStat(float m, float v, float x) {
   x_ = x;
 }
 
-float ProbabilityStat::precision() {
+double ProbabilityStat::precision() {
   return 1 / (std_ * std_);
 }
 
-float ProbabilityStat::prob() {
+double ProbabilityStat::prob() {
   if(prob_ == -1) {
     prob_ = dnorm(x_, mean_, std_);
   }
   return prob_;
 }
 
-float ProbabilityStat::dist_mean() {
+double ProbabilityStat::dist_mean() {
   return fabs((x_ - mean_) / std_);
 }
